Split renderer to it’s own sub-package.
